backend/requirements.txt
Flask
PyJWT
facepy
datetime
Flask_cors
pymongo
bson
dnspython
werkzeug
boto3